# 8261847: performance of java.lang.Record::toString should be improved