… test/langtools/tools/javac/records/BigRecordsToStringTest.java