Please review this PR which aims to optimize the implementation of the `toString` method we provide for records. A benchmark comparing the implementation we are providing for records with lombok found out that lombok is much faster mainly because our implementation uses `String::format`. This fix is basically delegating on StringConcatFactory::makeConcatWithConstants which is faster.

TIA

This is the result of the benchmark comparing records to lombok with vanilla JDK:
```
Benchmark                          Mode  Cnt    Score    Error  Units
MyBenchmark.base                   avgt    4    0.849 ±  0.111  ns/op
MyBenchmark.equals_record          avgt    4    7.343 ±  2.740  ns/op
MyBenchmark.equals_value           avgt    4    6.644 ±  1.920  ns/op
MyBenchmark.record_hash_code       avgt    4    5.763 ±  3.882  ns/op
MyBenchmark.record_to_string       avgt    4  262.626 ± 12.574  ns/op               <------ Before
MyBenchmark.value_class_to_string  avgt    4   30.325 ± 21.389  ns/op
MyBenchmark.value_hash_code        avgt    4    5.048 ±  3.936  ns/op
```

after this patch:
```
Benchmark                          Mode  Cnt   Score   Error  Units
MyBenchmark.base                   avgt    4   0.680 ± 0.185  ns/op
MyBenchmark.equals_record          avgt    4   5.599 ± 1.348  ns/op
MyBenchmark.equals_value           avgt    4   5.718 ± 4.633  ns/op
MyBenchmark.record_hash_code       avgt    4   4.628 ± 4.368  ns/op
MyBenchmark.record_to_string       avgt    4  26.791 ± 1.817  ns/op                 <------- After
MyBenchmark.value_class_to_string  avgt    4  35.473 ± 2.626  ns/op
MyBenchmark.value_hash_code        avgt    4   6.152 ± 5.101  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8261847](https://bugs.openjdk.java.net/browse/JDK-8261847): performance of java.lang.Record::toString should be improved


### Reviewers
 * [Jim Laskey](https://openjdk.java.net/census#jlaskey) (@JimLaskey - **Reviewer**) ⚠️ Review applies to 0ccb37abd1c59c1f82c4c96d6816b96b80bee2dc
 * [Claes Redestad](https://openjdk.java.net/census#redestad) (@cl4es - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/6403/head:pull/6403` \
`$ git checkout pull/6403`

Update a local copy of the PR: \
`$ git checkout pull/6403` \
`$ git pull https://git.openjdk.java.net/jdk pull/6403/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 6403`

View PR using the GUI difftool: \
`$ git pr show -t 6403`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/6403.diff">https://git.openjdk.java.net/jdk/pull/6403.diff</a>

</details>