@JimLaskey 2021-11-16
As Aleksey mentioned, you need a fallback for more that 200 components (limit of StringConcatFactory).