@cl4es 2021-11-16
More specifically the `StringConcatFactory` limit is 200 argument slots, not arguments.

Tests exercising maxed out records sounds like something we need to add if such doesn't already exist somewhere.