@vicente-romero-oracle 2021-11-17
I have uploaded another iteration, some numbers for when the number of record components is pushed to the maximum (254 components):

JDK18 vanilla:
```
Before:
Benchmark                          Mode  Cnt      Score      Error  Units
MyBenchmark.base                   avgt    4      0.670 ±    0.078  ns/op
MyBenchmark.equals_record          avgt    4   9092.649 ±  400.200  ns/op
MyBenchmark.equals_value           avgt    4    107.184 ±   15.993  ns/op
MyBenchmark.record_hash_code       avgt    4   6284.711 ± 1985.212  ns/op
MyBenchmark.record_to_string       avgt    4  69614.364 ± 6310.593  ns/op
MyBenchmark.value_class_to_string  avgt    4  25910.041 ± 7822.274  ns/op
MyBenchmark.value_hash_code        avgt    4    315.969 ±   36.545  ns/op
```

With this PR:
```
Benchmark                          Mode  Cnt      Score      Error  Units
MyBenchmark.base                   avgt    4      0.658 ±    0.053  ns/op
MyBenchmark.equals_record          avgt    4   8222.407 ±  308.440  ns/op
MyBenchmark.equals_value           avgt    4    102.627 ±   56.522  ns/op
MyBenchmark.record_hash_code       avgt    4   6030.478 ±  243.390  ns/op
MyBenchmark.record_to_string       avgt    4  46504.151 ± 3597.182  ns/op
MyBenchmark.value_class_to_string  avgt    4  26196.565 ± 8154.037  ns/op
MyBenchmark.value_hash_code        avgt    4    313.918 ±    4.577  ns/op
```

some observations, we are doing better than before but still worst than lombok. I don't think having to create a more complicated MH tree is the key, our numbers are still worst even if there is no need to create too much levels in the MH tree like for 100 record components occupying a spot each. Another observation is that even though our slowest implementation is the one for the `toString` method, in absolute terms, lombok is faster in all cases. So more work could be necessary for the rest of the methods if deemed worthy.

Thanks!