@vicente-romero-oracle 2021-11-19
I have done some additional experiments as suggested by Claes, thanks, for different number of record components. In all cases all the components are of type `int` so they occupy only one slot. Here are some numbers. I tried with 0, 1, 10, 100 and 254 record components.
```
Benchmark                                  Mode  Cnt      Score       Error  Units
MyBenchmark.base                           avgt    4      0.775 ±     0.552  ns/op
MyBenchmark.record0_toString               avgt    4      4.973 ±     2.800  ns/op
MyBenchmark.record1_toString               avgt    4     16.026 ±     5.100  ns/op
MyBenchmark.record10_toString              avgt    4     81.412 ±     3.365  ns/op
MyBenchmark.record100_toString             avgt    4  12269.500 ±   179.784  ns/op
MyBenchmark.record254_toString             avgt    4  51191.953 ± 11679.762  ns/op
MyBenchmark.valueClass0_toString           avgt    4      5.134 ±     2.372  ns/op
MyBenchmark.valueClass1_toString           avgt    4     23.321 ±     9.234  ns/op
MyBenchmark.valueClass10_toString          avgt    4     94.048 ±     7.017  ns/op
MyBenchmark.valueClass100_toString         avgt    4   9253.282 ±  4843.738  ns/op
MyBenchmark.valueClass254_toString         avgt    4  31963.158 ± 24050.499  ns/op
```

Then also after a suggestion from Claes I modified the maximum number of slots I would be chopping the arguments into, I gave a try with 20 slots and got these numbers:

```
Benchmark                                  Mode  Cnt      Score       Error  Units
MyBenchmark.record0_toString               avgt    4      5.009 ±     3.454  ns/op
MyBenchmark.record1_toString               avgt    4     14.207 ±    10.551  ns/op
MyBenchmark.record10_toString              avgt    4     81.018 ±     7.320  ns/op
MyBenchmark.record100_toString             avgt    4   2862.641 ±  1233.862  ns/op
MyBenchmark.record254_toString             avgt    4  23002.280 ± 97103.923  ns/op
MyBenchmark.valueClass0_toString           avgt    4      4.967 ±     3.947  ns/op
MyBenchmark.valueClass1_toString           avgt    4     23.756 ±     8.499  ns/op
MyBenchmark.valueClass10_toString          avgt    4     87.691 ±     7.956  ns/op
MyBenchmark.valueClass100_toString         avgt    4   9539.272 ±  9461.516  ns/op
MyBenchmark.valueClass254_toString         avgt    4  28323.478 ± 11932.474  ns/op
```
It seems like the execution is way faster now.